+++ omniva-woocommerce/assets/blocks/terminal-selection-block/cart/index.asset.php
@@ -1,1 +1,1 @@
-<?php return array('dependencies' => array('react', 'wp-block-editor', 'wp-blocks', 'wp-components', 'wp-i18n'), 'version' => '258e2ab873a31bf52b06');
+<?php return array('dependencies' => array('react', 'wp-block-editor', 'wp-blocks', 'wp-components', 'wp-i18n'), 'version' => '578f2e6b6f30bf94d1e3');